Add a service container.
Add the various pieces that will be needed for the watch functions list
to the service container. Also add a stub for the watch controller as
well.

Using pimple will allow the controller objects and all their
dependencies to be separated from the global soup that Slim naturally
encourages. My goal is to have functional tests for all the various
controller methods.