Add a configuration hook to customizing how URLs are generated.
Allow a callback in the configuration that makes it easier to customize
how the simple_url is generated. The simple_url value is used to
identify 'unique' routes or blocks of application logic.

Keep the existing default of trimming out numeric GET parameters, so
people upgrading don't get wildly different URL's than before.