Move MongoDB dependency from require to suggest
Since XHGui is often used on the production servers to gather
the profiling data, it is inconvenient that it requires MongoDB
for the collection part as well, when Mongo is actually necessary
only for displaying the data.

The profiling workflow in a production environment should be as follows:
1) Include XHGui as a dependency for the application.
2) Gather profiling data.
3) Save the results into a file (on the production machine).
4) Download the file to a local dev machine.
5) Import the file into Mongo.
6) Display the results (on a local dev machine).

This commit ensures that XHGui can be installed without MongoDB
as a dependency.

This has also been discussed in issue #85