Add more getFlamegraphData tests
This ensures a test exists for cases where there are multiple
call stacks that have the same function in their leaf position.

This is a known use case that XHGui accounts with its calculateSelf()
method, but it seems the current getFlameGraphData() method uses
this data incorrectly.

This issue is mentioned at https://github.com/perftools/xhgui/pull/216#issuecomment-344154508

Before I fix this, I'm first adding a test to show the current broken
behaviour.