Use non-aggregated data when generating the flamegraph.
This will help reduce the number of flame towers with wide blocks on top
of narrow ones, allowing us to upgrade d3.flamegraph more easily.

We can't entirely remove the wide blocks on top of narrow ones, as
xhprof does not expose enough data for deeply nested call stacks. We
could apply heuristics to ensure that a child is always lesser than its
parent, but we will report data inaccurately with that approach.