Update README.md with current compatibility info
Update the XHGUI range to include 0.8 and 0.9 to clarify that
there still have not been backwards-incompatible schema changes.

Keep using explicit version number for XHGUI because may break
compatibility in a 0.x release.

Also update the collector version start explicitly at 1.0.0 to indicate that as of writing all
versions have been cross-compatible. Keep the collector's range as `1.x` instead of
explicit because it is not expected for a future `1.x` release to break compatibility.

Fixes #9.