require php 5.6+, add testing upto 7.2, upgrade phpunit
The use of namespaces (new in PHPUnit 6) was backported to
PHPUnit 5, so we can use both PHPUnit 5 (PHP 5.6-7.1) and
PHPUnit 6 (PHP 7-7.2) on the same test suites.

PHP 5.5 went EOL in 2016,
<https://php.net/eol>,
<https://php.net/supported-versions>.
And PHPUnit 4 went EOL in 2017,
<https://phpunit.de/supported-versions.html>.

PHP 5.6 got an extra year of extended support, but also went EOL
as of December 31. However, I'd like to move to PHP 7+ in a
separate commit for the time being, but we could do both in the
same release.

The impact would be that those using LTS releases and Debian Linux,
that are on Debian 8 Jessie (still supported), which comes with
php 5.6, couldn't upgrade to latest master until they upgrade to
Debian 9 Stretch (php7.0), or use backports.

Ubunto 16 Xenial comes with php7.0, and Ubuntu 18 Bionic, php7.2.